fix: refactor k8s & k8swatcher
