+++ bases/renku_data_services/data_api/dependencies.py
@@ -48,8 +48,8 @@ from renku_data_services.k8s.clients import (
     K8sSchedulingClient,
 )
 from renku_data_services.k8s.config import KubeConfigEnv
+from renku_data_services.k8s.db import K8sDbCache
 from renku_data_services.k8s.quota import QuotaRepository
-from renku_data_services.k8s_watcher import K8sDbCache
 from renku_data_services.message_queue.db import ReprovisioningRepository
 from renku_data_services.metrics.core import StagingMetricsService
 from renku_data_services.metrics.db import MetricsRepository
