+++ components/renku_data_services/k8s/db.py
@@ -11,7 +11,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from renku_data_services.errors import errors
 from renku_data_services.k8s.models import K8sObject, K8sObjectFilter, K8sObjectMeta
-from renku_data_services.k8s_watcher.orm import K8sObjectORM
+from renku_data_services.k8s.orm import K8sObjectORM
 
 
 class K8sDbCache:
